+++ R/simulate_infections.R
@@ -33,6 +33,7 @@
 #'   growth roughly in line with the growth rate implied by the first value of
 #'   R.
 #' @inheritParams estimate_infections
+#' @inheritParams calc_CrIs
 #' @inheritParams rt_opts
 #' @inheritParams stan_opts
 #' @importFrom checkmate assert_data_frame assert_date assert_numeric
@@ -74,12 +75,11 @@ simulate_infections <- function(R,
                                 growth_method = c("infections",
                                                   "infectiousness")) {
   if (is.numeric(pop)) {
-    lifecycle::deprecate_warn(
-      "1.7.0",
+    lifecycle::deprecate_stop(
+      "1.9.0",
       "simulate_infections(pop = 'must be a `<dist_spec>`')",
       details = "For specifying a fixed population size, use `Fixed(pop)`"
     )
-    pop <- Fixed(pop)
   }
   pop_period <- arg_match(pop_period)
   if (pop_period == "all" && pop == Fixed(0)) {